requirements.txt
flask
pillow
opencv-python
torch
torchvision
tensorflow
ultralytics
numpy
matplotlib
boto3
pymongo